requirements.txt
torch
torchvision
numpy
pandas
matplotlib
captum
streamlit
pillow
pyngrok
tqdm